requirements.txt
# Core
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.1,<1.0.0
gunicorn>=21.2.0,<22.0.0
starlette>=0.36.3,<1.0.0

# Database
SQLAlchemy[asyncio]>=2.0.29,<3.0.0
alembic>=1.13.1,<2.0.0
asyncpg>=0.29.0,<1.0.0
# greenlet==3.0.3  # disabled due to build issues on Python 3.13
psycopg2-binary>=2.9.9,<3.0.0

# Pydantic
pydantic-settings>=2.2.1,<3.0.0
pydantic[email]>=2.6.4,<3.0.0

# HTTP
httpx>=0.27.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
aiohttp>=3.9.3,<4.0.0
backoff>=2.2.1,<3.0.0

# Auth & Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=3.2.0,<4.0.0
cryptography>=42.0.5,<43.0.0

# Rate Limiting
fastapi-limiter>=0.1.6,<1.0.0
redis>=5.0.1,<6.0.0

# Validation
email-validator>=2.1.0,<3.0.0

# OpenTelemetry Core
opentelemetry-api>=1.24.0,<2.0.0
opentelemetry-sdk>=1.24.0,<2.0.0
opentelemetry-semantic-conventions>=0.45b0,<1.0.0
opentelemetry-util-http>=0.45b0,<1.0.0

# OpenTelemetry Instrumentation
opentelemetry-instrumentation>=0.45b0,<1.0.0
opentelemetry-instrumentation-fastapi>=0.45b0,<1.0.0
opentelemetry-instrumentation-sqlalchemy>=0.45b0,<1.0.0
opentelemetry-instrumentation-redis>=0.45b0,<1.0.0
opentelemetry-instrumentation-requests>=0.45b0,<1.0.0
opentelemetry-instrumentation-logging>=0.45b0,<1.0.0
opentelemetry-instrumentation-aiohttp-client>=0.45b0,<1.0.0
opentelemetry-instrumentation-asgi>=0.45b0,<1.0.0
opentelemetry-instrumentation-asyncpg>=0.45b0,<1.0.0
opentelemetry-instrumentation-psycopg2>=0.45b0,<1.0.0

# OpenTelemetry Exporters
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-prometheus==0.45b0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-proto==1.24.0

# Metrics and Monitoring
prometheus-client==0.20.0
sentry-sdk[fastapi]==1.39.1

# Email Service
aiosmtplib>=3.0.0,<4.0.0
jinja2>=3.1.0,<4.0.0

# Background tasks
celery[redis]>=5.3.6,<6.0.0
nest-asyncio>=1.5.8,<2.0.0

# Structured Logging
structlog==24.2.0
python-json-logger==3.0.1
tenacity==8.2.3
pytz>=2023.3

# Data Processing
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
xlrd>=2.0.0,<3.0.0
Pillow>=10.0.0,<11.0.0
requests>=2.31.0,<3.0.0

# Chinese Text Processing
jieba>=0.42.1,<1.0.0

# Google Sheets Integration
gspread>=5.12.0,<6.0.0
oauth2client>=4.1.3,<5.0.0

# Testing
pytest>=8.1.1,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0

# Dependency Injection
dependency-injector>=4.41.0,<5.0.0
aiosqlite>=0.19.0,<1.0.0
